move everything to llama_ring
